backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
psycopg2-binary==2.9.10
pandas==2.3.0
numpy>=1.24.0
python-dotenv==1.0.1
